debugging output and conditional killing

+++ runbench.sh
@@ -10,11 +10,14 @@ HOSTS_FILE=$1
 MONGOD_SERVER=$(head -n 1 $HOSTS_FILE)
 
 tail -n +2 $HOSTS_FILE | while read host; do
+	echo "$host -> $MONGOD_SERVER"
 	ssh $host "~/mongo/perfbench/rampandhold.sh $MONGOD_SERVER"
 	scp ${host}:testinfo.json $host-testinfo.json
 	scp ${host}:holdit.pid $host-holdit.pid
 done
 
 tail -n +2 $HOSTS_FILE | while read host; do
-	ssh $host "kill $(cat $host-holdit.pid)" 
+	if [ -f $host-holdit.pid ]; then
+		ssh $host "kill $(cat $host-holdit.pid)" 
+	fi
 done
